Rename openapi-decorators to openapi-metadata
