+++ packages/openapi-adonis/src/decorators/index.ts
@@ -1,1 +1,1 @@
-export * from "openapi-decorators";
+export * from "openapi-metadata";
