+++ packages/openapi-metadata/.npmignore
