+++ packages/openapi-metadata/src/builders/document-builder.ts
