+++ packages/openapi-metadata/src/fixtures/document-base.ts
