+++ packages/openapi-metadata/src/types.ts
