+++ packages/openapi-metadata/src/ui/scalar.ts
