+++ packages/openapi-metadata/src/utils/metadata.ts
